moved the login and signup fragments to user directory
